add antialiasing to point drawing
Implement antialising by drawing subpixel-positioned points as
downsampling kernels. I implemented two kernels: one based on bilinear
interpolation, and a lanczos kernel.

They work reasonably well for grayscale images, but because the colormap is
being applied after the antialiasing, it creates some artifacts in
antialised curves (a supossed white curve would have a yellow/red mixed
in).

A better aproach (although maybe slower) may be to render the attractor
to a bigger aggregation buffer, and downsample that in the fragment
shader, after applying the colormap.