implementation parallel rendering
It is only a little faster than the singlethreaded version. Previous
experiments have show that rendering to a Atomic buffer is around 2
times slower than a non-atomic buffer. Because my machine has only 4
cores, the results are reasonable. Maybe with more cores it will be much
faster.

Another approach is to render the multiple buffers and them aggregate
them together. This will cost much more memory, and there is the penalty
for summing them, but may be faster. I will test that next.