implement "merge" approach to multithreading
In this mode, each thread gains a clean copy of the buffer, and add the
end the copies are merged. This is faster than the atomic approach when
the number of samples per iteration is big enough.

Also added a control for the number of samples per iteration to the GUI.